+++ src/org/schema/enumerations/status/PaymentAutomaticallyApplied.php
@@ -2,13 +2,11 @@
 
 namespace org\schema\enumerations\status;
 
-use org\schema\enumerations\StatusEnumeration;
-
 /**
  * An automatic payment system is in place and will be used.
  * @see https://schema.org/PaymentAutomaticallyApplied
  */
-class PaymentAutomaticallyApplied extends StatusEnumeration
+class PaymentAutomaticallyApplied extends PaymentStatusType
 {
 
 }