+++ src/org/schema/enumerations/status/PaymentComplete.php
@@ -2,13 +2,11 @@
 
 namespace org\schema\enumerations\status;
 
-use org\schema\enumerations\StatusEnumeration;
-
 /**
  * The payment has been received and processed.
  * @see https://schema.org/PaymentComplete
  */
-class PaymentComplete extends StatusEnumeration
+class PaymentComplete extends PaymentStatusType
 {
 
 }