+++ src/org/schema/enumerations/status/PaymentDeclined.php
@@ -2,13 +2,11 @@
 
 namespace org\schema\enumerations\status;
 
-use org\schema\enumerations\StatusEnumeration;
-
 /**
  * The payee received the payment, but it was declined for some reason.
  * @see https://schema.org/PaymentDeclined
  */
-class PaymentDeclined extends StatusEnumeration
+class PaymentDeclined extends PaymentStatusType
 {
 
 }