+++ src/org/schema/enumerations/status/PaymentDue.php
@@ -2,13 +2,11 @@
 
 namespace org\schema\enumerations\status;
 
-use org\schema\enumerations\StatusEnumeration;
-
 /**
  * The payment is due, but still within an acceptable time to be received.
  * @see https://schema.org/PaymentDue
  */
-class PaymentDue extends StatusEnumeration
+class PaymentDue extends PaymentStatusType
 {
 
 }