+++ src/org/schema/enumerations/status/PaymentPastDue.php
@@ -2,13 +2,11 @@
 
 namespace org\schema\enumerations\status;
 
-use org\schema\enumerations\StatusEnumeration;
-
 /**
  * The payment is due and considered late.
  * @see https://schema.org/PaymentPastDue
  */
-class PaymentPastDue extends StatusEnumeration
+class PaymentPastDue extends PaymentStatusType
 {
 
 }